Move gtag-wrapper.ts and universal-analytics-wrapper.ts into library-wrapper directory.
