+++ src-shared/analytics/analytics-main.ts
@@ -1,7 +1,7 @@
 import { GoogleAnalytics4IpcMain } from './ipc/google-analytics-4-ipc';
 import { UniversalAnalyticsIpcMain } from './ipc/universal-analytics-ipc';
+import { UniversalAnalyticsWrapper } from './library-wrapper/universal-analytics-wrapper';
 import { AnalyticsInterface } from './analytics-interface';
-import { UniversalAnalyticsWrapper } from './universal-analytics-wrapper';
 
 export class AnalyticsMain implements AnalyticsInterface {
   constructor() {
